requirements.txt
python-telegram-bot==20.0a6
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openai
transformers
torch
